Final Frontend refactoring
