+++ frontend/src/components/Skeleton/components/InstalledAppSearch.jsx
