+++ frontend/src/components/Skeleton/components/ToolsSections.jsx
