--- frontend/src/components/searchbar.component1.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SearchBar = () => {
-  return <div className="text-center">This is a a search bar</div>;
-};
-
-export default SearchBar;
