+++ frontend/src/components/utils/data/requiredData.jsx
